[#3360] make sure @Required gets applied to all properties
