+++ sormas-rest/src/main/java/de/symeda/sormas/rest/swagger/SwaggerConfig.java
@@ -17,69 +17,59 @@
  *******************************************************************************/
 package de.symeda.sormas.rest.swagger;
 
-import java.util.ArrayList;
-import java.util.List;
-
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
-
 import io.swagger.v3.core.converter.ModelConverters;
 import io.swagger.v3.core.jackson.ModelResolver;
 import io.swagger.v3.core.util.Json;
-import io.swagger.v3.jaxrs2.ext.OpenAPIExtension;
 import io.swagger.v3.jaxrs2.ext.OpenAPIExtensions;
 
 /**
  * SORMAS Swagger Configuration.
  * This class is the preferred spot for specifying API-related Swagger metadata via Swagger annotations,
  * as well as for Swagger extension registration.
- *
- * @author Jan-Niklas Brandes
  */
 public class SwaggerConfig {
 
-	static {
-		// Real initialization routine
-		registerExtensions();
+    static {
+        // Real initialization routine
+        registerExtensions();
 
-		// Swagger uses a Jackson ObjectMapper in the process of type resolution; there are some
-		// settings for that ObjectMapper and Swagger-related classes we need to adjust for the
-		// Swagger Specification to be correct
-		tweakConfig(Json.mapper());
-	}
+        // Swagger uses a Jackson ObjectMapper in the process of type resolution; there are some
+        // settings for that ObjectMapper and Swagger-related classes we need to adjust for the
+        // Swagger Specification to be correct
+        tweakConfig(Json.mapper());
+    }
 
-	public static void init() {
-		// Solemnly there for triggering the static initializer
-	}
+    public static void init() {
+        // Solemnly there for triggering the static initializer
+    }
 
-	/**
-	 * Register Swagger Generator Extensions
-	 */
-	private static void registerExtensions() {
+    /**
+     * Register Swagger Generator Extensions
+     */
+    private static void registerExtensions() {
+        // rewrite controller strings
+        OpenAPIExtensions.getExtensions().add(new SormasSwaggerExtensions());
 
-		SormasSwaggerExtensions sormasSwaggerExtension = new SormasSwaggerExtensions();
+        // include annotations in openAPI specs
+        ModelConverters.getInstance().addConverter(new AttributeConverter(Json.mapper()));
+    }
 
-		// Operations-Level extensions
-		List<OpenAPIExtension> swaggerExtensions = new ArrayList<>(OpenAPIExtensions.getExtensions());
-		swaggerExtensions.add(sormasSwaggerExtension);
+    /**
+     * Set configuration parameters for Swagger-related classes.
+     *
+     * @param swaggerObjectMapper ObjectMapper instance used by Swagger
+     */
+    private static void tweakConfig(ObjectMapper swaggerObjectMapper) {
+        // Do not use toString() on enum values
+        swaggerObjectMapper.disable(SerializationFeature.WRITE_ENUMS_USING_TO_STRING);
 
-		OpenAPIExtensions.setExtensions(swaggerExtensions);
+        // Specify enumerations as separate schemas, instead of incorporating them into the
+        // schemas for every field/property of their type
+        ModelResolver.enumsAsRef = true;
+    }
 
-		// Schema-Level extensions
-		ModelConverters modelConverters = ModelConverters.getInstance();
-		modelConverters.addConverter(sormasSwaggerExtension);
-	}
-
-	/**
-	 * Set configuration parameters for Swagger-related classes.
-	 * @param swaggerObjectMapper ObjectMapper instance used by Swagger
-	 */
-	private static void tweakConfig(ObjectMapper swaggerObjectMapper) {
-		// Do not use toString() on enum values
-		swaggerObjectMapper.disable(SerializationFeature.WRITE_ENUMS_USING_TO_STRING);
-
-		// Specify enumerations as separate schemas, instead of incorporating them into the
-		// schemas for every field/property of their type
-		ModelResolver.enumsAsRef = true;
-	}
 }
+
+
